requirements.txt
pymupdf
python-docx
nltk
spacy
scikit-learn
sentence-transformers
networkx
faiss-cpu
transformers
torch
fastapi
uvicorn
sqlite-utils
psutil
matplotlib
numpy
pandas

